Moved page components to src/pages
