+++ src/pages/NQueenPuzzle/NQueenPuzzle.css
