+++ src/pages/NQueenPuzzle/NQueenPuzzle.test.js
